requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
Flask-Mail==0.10.0
fonttools==4.52.4
frozendict==2.4.4
html5lib==1.1
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pandas-datareader==0.10.0
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
retrying==1.3.4
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
tenacity==8.3.0
threadpoolctl==3.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.40
zipp==3.19.0
